+++ nconfigs/libs/libtestfr.c
@@ -14,5 +14,17 @@ VMValue unsigned_add(VMValue* args, uint32_t argc) {
     uint64_t b = args[1].data;
 
     uint64_t res = a + b;
+    printf("From lib: %ld", res);
+    return (VMValue){.typeind=1, .data=res};
+}
+
+VMValue unsigned_pow2(VMValue* args, uint32_t argc) {
+    if (argc < 1) {
+        return (VMValue){.typeind=1, .data=0}; // 1 - uint
+    }
+    uint64_t a = args[0].data;
+
+    uint64_t res = a * a;
+    printf("From lib: %ld", res);
     return (VMValue){.typeind=1, .data=res};
 }