+++ changelog.md
@@ -2,3 +2,11 @@ CHANGELOG v8:
 pre1:
 - operations gsf (), usf () for getting/updating stack frames respectively
 - native interface, see src/native.rs
+pre2:
+- major refactor of registers
+pre3:
+- much faster init now - just changed `system::new_all` -> `system::new`!
+pre4:
+- better ffi config
+- some stuff i forgot
+todo: todo.md